Restructure the implementation of cluster support
